Fix css loading

+++ Migrations/V011__Stats.sql
@@ -4,7 +4,7 @@ OR REPLACE FUNCTION get_user_top_categories (p_user_id INT) RETURNS TABLE (
     category_id INT,
     category_name VARCHAR(32),
     accuracy_rate NUMERIC,
-    total_questions INT
+    total_questions BIGINT
 ) AS $$
 BEGIN
     RETURN QUERY
